requirements/common.txt
django == 5.0.6
mysqlclient==2.2.4
pillow==10.4.0
django-ajax-datatable==4.5.0
django-crispy-forms==2.2
django-environ==0.11.2
crispy-bootstrap5==2024.2
dj-database-url==2.2.0